Add interfaces for responses from PokeApi
Interfaces for the data from PokeApi have been create for each endpoint
as described in http://www.pokeapi.co/docsv2/. Currently every interface
is exported and some of them are not exactly where they should be. The
organization, naming, and interfaces themselves are subject to change in
the future as needed.

Common models for data from the API are described in common.ts. These
include interfaces like Resource, Name, Description, and other simple
data models.

sprites.ts stores all the sprite interfaces that are used throughout the
API.

Every major section is put into its own folder, so berry endpoints are
in one folder, evolution endpoints are in another, and so on.