Create script to cache some results locally
This script takes a very long time to run, usually taking 3ish hours.
Only run this script when major changes to the API occur.

Store simple information for pokemon for the sake of detailed list
views. In order to do this without requesting this info all up front
would be to make several hundreds of requests when needed, which is not
feasible.